requirements.txt
# ============================================================================
# Dagster ETL - Requirements
# ============================================================================
# Python 3.12+

# Dagster Core
dagster>=1.9.0
dagster-webserver>=1.9.0
dagster-postgres>=0.27.0
dagster-dbt>=0.28.0

# dbt
dbt-core>=1.8.0
dbt-postgres>=1.8.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.24.0

# Prefect (pour compatibilité code existant)
prefect>=3.0.0

# Utilities
python-dotenv>=1.0.0
tenacity>=9.0.0
requests>=2.31.0